image_types: Remove leading './' in tar image archives
The leading './' makes it difficult to provide relative paths to tar
during the untar operation.
For example, `tar xvf --exclude=usr/bin` does not apply correctly to
the files in the archive because they are prefixed with a './'. As an
alternative to applying './' to all the exclude paths, remove the
leading './' when creating the tar archive.

Signed-off-by: Chaitanya Vadrevu <chaitanya.vadrevu@ni.com>
Signed-off-by: Shruthi Ravichandran <shruthi.ravichandran@ni.com>

Rebased to kirkstone and updated variable override syntax.

Rebased-by: Alex Stewart <alex.stewart@ni.com>